refactor: Q&A 관련 내부 DTO를 inner record로 변경
