+++ backend/ddang/src/main/java/com/ddang/ddang/qna/application/AnswerService.java
@@ -1,7 +1,7 @@
 package com.ddang.ddang.qna.application;
 
 import com.ddang.ddang.auction.application.exception.UserForbiddenException;
-import com.ddang.ddang.qna.application.dto.CreateAnswerDto;
+import com.ddang.ddang.qna.application.dto.request.CreateAnswerDto;
 import com.ddang.ddang.qna.application.event.AnswerNotificationEvent;
 import com.ddang.ddang.qna.application.exception.AlreadyAnsweredException;
 import com.ddang.ddang.qna.application.exception.InvalidAnswererException;
