+++ backend/ddang/src/main/java/com/ddang/ddang/qna/presentation/QnaController.java
@@ -6,8 +6,8 @@ import com.ddang.ddang.image.presentation.util.ImageRelativeUrlFinder;
 import com.ddang.ddang.image.presentation.util.ImageTargetType;
 import com.ddang.ddang.qna.application.AnswerService;
 import com.ddang.ddang.qna.application.QuestionService;
-import com.ddang.ddang.qna.application.dto.CreateAnswerDto;
-import com.ddang.ddang.qna.application.dto.CreateQuestionDto;
+import com.ddang.ddang.qna.application.dto.request.CreateAnswerDto;
+import com.ddang.ddang.qna.application.dto.request.CreateQuestionDto;
 import com.ddang.ddang.qna.presentation.dto.request.CreateAnswerRequest;
 import com.ddang.ddang.qna.presentation.dto.request.CreateQuestionRequest;
 import jakarta.validation.Valid;
