+++ backend/ddang/src/test/java/com/ddang/ddang/qna/presentation/QnaControllerTest.java
@@ -31,8 +31,8 @@ import com.ddang.ddang.authentication.domain.TokenDecoder;
 import com.ddang.ddang.authentication.domain.TokenType;
 import com.ddang.ddang.authentication.domain.dto.AuthenticationStore;
 import com.ddang.ddang.exception.GlobalExceptionHandler;
-import com.ddang.ddang.qna.application.dto.CreateAnswerDto;
-import com.ddang.ddang.qna.application.dto.CreateQuestionDto;
+import com.ddang.ddang.qna.application.dto.request.CreateAnswerDto;
+import com.ddang.ddang.qna.application.dto.request.CreateQuestionDto;
 import com.ddang.ddang.qna.application.exception.AlreadyAnsweredException;
 import com.ddang.ddang.qna.application.exception.InvalidAnswererException;
 import com.ddang.ddang.qna.application.exception.InvalidAuctionToAskQuestionException;
